+++ lib/actions/upsertPerson.js
@@ -18,10 +18,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-const Q = require('q');
-const request = require('request-promise');
-const { messages } = require('elasticio-node');
-const { createSession } = require('./../utils/wice');
+const { upsertObject, newMessage } = require('./../utils/helpers');
 
 /**
  * This method will be called from elastic.io platform providing following data
@@ -30,127 +27,37 @@ const { createSession } = require('./../utils/wice');
  * @param cfg configuration that is account information and configuration field values
  */
 
-async function checkForExistingPerson(person, cookie, options) {
-  let existingRowid = 0;
-  try {
-    options.form = {
-      method: 'get_person',
-      cookie,
-      pkey: person.body.meta.recordUid,
-    };
-
-
-    const rowid = await request.post(options);
-
-    const rowidObj = JSON.parse(rowid);
-    if (rowidObj.address_object) {
-      existingRowid = rowidObj.address_object.rowid;
-      console.log(`Person already exists ... ROWID: ${existingRowid}`);
-    }
-    return existingRowid;
-  } catch (e) {
-    throw new Error(e);
-  }
-}
-
-async function createOrUpdatePerson(existingRowid, cookie, options, msg) {
-  try {
-    if (existingRowid === 0) {
-      console.log('Creating person ...');
-      // const input = JSON.stringify(msg.body);
-      const input = JSON.stringify(msg.body.data);
-      options.form = {
-        method: 'insert_contact',
-        data: input,
-        cookie,
-      };
-      const person = await request.post(options);
-      return JSON.parse(person);
-    }
-    console.log('Updating person ...');
-    // msg.body.rowid = existingRowid;
-    msg.body.data.rowid = existingRowid;
-    options.form = {
-      method: 'update_contact',
-      // data: JSON.stringify(msg.body),
-      data: JSON.stringify(msg.body.data),
-      cookie,
-    };
-
-    const person = await request.post(options);
-    return JSON.parse(person);
-  } catch (e) {
-    throw new Error(e);
-  }
-}
-
-async function executeRequest(msg, cfg, options) {
-  try {
-    const cookie = await createSession(cfg);
-    const existingRowid = await checkForExistingPerson(msg, cookie, options);
-    const result = await createOrUpdatePerson(existingRowid, cookie, options, msg);
-    return result;
-  } catch (e) {
-    throw new Error(e);
-  }
-}
-
 async function processAction(msg, cfg) {
-  const self = this;
-  const oihUid = (msg.body.meta !== undefined && msg.body.meta.oihUid !== undefined) ? msg.body.meta.oihUid : 'oihUid not set yet';
-  const recordUid = (msg.body !== undefined && msg.body.meta.recordUid !== undefined) ? msg.body.meta.recordUid : undefined;
-  const applicationUid = (msg.body.meta !== undefined && msg.body.meta.applicationUid !== undefined) ? msg.body.meta.applicationUid : undefined;
+  try {
+    const oihUid = (msg.body.meta !== undefined && msg.body.meta.oihUid !== undefined) ? msg.body.meta.oihUid : undefined;
+    const recordUid = (msg.body !== undefined && msg.body.meta.recordUid !== undefined) ? msg.body.meta.recordUid : undefined;
+    const applicationUid = (msg.body.meta !== undefined && msg.body.meta.applicationUid !== undefined) ? msg.body.meta.applicationUid : undefined;
 
-  /** The following block creates the meta object.
+    /** The following block creates the meta object.
    *  This meta object stores information which are
    *  later needed in order to make the hub and spoke architecture work properly
    */
-  const oihMeta = {
-    applicationUid,
-    oihUid,
-    recordUid,
-  };
+    const oihMeta = {
+      applicationUid,
+      oihUid,
+      recordUid,
+    };
 
-  const options = {
-    method: 'POST',
-    uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json`,
-    headers: {
-      'X-API-KEY': cfg.apikey,
-    },
-  };
-
-  msg.body.data.same_contactperson = 'auto';
-
-  async function emitData() {
-    const reply = await executeRequest(msg, cfg, options);
+    const reply = await upsertObject(msg, cfg, 'person');
     oihMeta.recordUid = reply.rowid;
     const response = {
       meta: oihMeta,
       data: reply,
     };
 
-    const data = messages.newMessageWithBody(response);
-    self.emit('data', data);
-  }
-
-  function emitError(e) {
+    const data = newMessage(response);
+    this.emit('data', data);
+  } catch (e) {
     console.log('Oops! Error occurred');
-    self.emit('error', e);
+    this.emit('error', e);
   }
-
-  function emitEnd() {
-    console.log('Finished execution');
-    self.emit('end');
-  }
-
-  Q()
-    .then(emitData)
-    .fail(emitError)
-    .done(emitEnd);
 }
 
 module.exports = {
   process: processAction,
-  checkForExistingPerson,
-  createOrUpdatePerson,
 };
