+++ lib/utils/helpers.js
@@ -145,7 +145,144 @@ async function getOrganizations(cfg, snapshot) {
   }
 }
 
+async function checkForExistingPerson(person, cookie, options) {
+  let existingRowid = 0;
+  try {
+    const newOptions = options;
+    newOptions.form = {
+      method: 'get_person',
+      cookie,
+      pkey: person.body.meta.recordUid,
+    };
+
+
+    const rowid = await request.post(newOptions);
+
+    const rowidObj = JSON.parse(rowid);
+    if (rowidObj.address_object) {
+      existingRowid = rowidObj.address_object.rowid;
+      console.log(`Person already exists ... ROWID: ${existingRowid}`);
+    }
+    return existingRowid;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+async function createOrUpdatePerson(existingRowid, cookie, options, msg) {
+  try {
+    const newOptions = options;
+    const newMsg = msg;
+    if (existingRowid === 0) {
+      console.log('Creating person ...');
+
+      const input = JSON.stringify(msg.body.data);
+      newOptions.form = {
+        method: 'insert_contact',
+        data: input,
+        cookie,
+      };
+      const person = await request.post(newOptions);
+      return JSON.parse(person);
+    }
+    console.log('Updating person ...');
+    // msg.body.rowid = existingRowid;
+    newMsg.body.data.rowid = existingRowid;
+    newOptions.form = {
+      method: 'update_contact',
+      // data: JSON.stringify(msg.body),
+      data: JSON.stringify(msg.body.data),
+      cookie,
+    };
+
+    const person = await request.post(newOptions);
+    return JSON.parse(person);
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+async function checkForExistingOrganization(organization, cookie, options) {
+  let existingRowid = 0;
+  try {
+    const newOptions = options;
+    newOptions.form = {
+      method: 'get_company',
+      cookie,
+      pkey: organization.body.meta.recordUid,
+    };
+
+    const rowid = await request.post(newOptions);
+    const rowidObj = JSON.parse(rowid);
+    if (rowidObj.address_object) {
+      existingRowid = rowidObj.address_object.rowid;
+      console.log(`Organization already exists ... ROWID: ${existingRowid}`);
+    }
+    return existingRowid;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+async function createOrUpdateOrganization(existingRowid, cookie, options, msg) {
+  try {
+    const newOptions = options;
+    if (existingRowid === 0) {
+      console.log('Creating organization ...');
+      const input = JSON.stringify(msg.body.data);
+      newOptions.form = {
+        method: 'insert_company',
+        data: input,
+        cookie,
+      };
+      const organization = await request.post(newOptions);
+      return JSON.parse(organization);
+    }
+    const newMsg = msg;
+    console.log('Updating organization ...');
+    newMsg.body.data.rowid = existingRowid;
+    newOptions.form = {
+      method: 'update_company',
+      data: JSON.stringify(msg.body.data),
+      cookie,
+    };
+    const organization = await request.post(newOptions);
+    return JSON.parse(organization);
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+
+async function upsertObject(msg, cfg, type) {
+  try {
+    let result;
+    const options = {
+      method: 'POST',
+      uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json`,
+      headers: {
+        'X-API-KEY': cfg.apikey,
+      },
+    };
+
+    const cookie = await createSession(cfg);
+    if (type === 'person') {
+      const newMsg = msg;
+      newMsg.body.data.same_contactperson = 'auto';
+
+      const existingRowid = await checkForExistingPerson(newMsg, cookie, options);
+      result = await createOrUpdatePerson(existingRowid, cookie, options, msg);
+    } else {
+      const existingRowid = await checkForExistingOrganization(msg, cookie, options);
+      result = createOrUpdateOrganization(existingRowid, cookie, options, msg);
+    }
+    return result;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
 
 module.exports = {
-  createSession, newMessage, getPersons, getOrganizations,
+  createSession, newMessage, getPersons, getOrganizations, upsertObject,
 };
